# docs/openapi.yaml
openapi: 3.0.0
info:
  title: Web Appliance Dashboard API
  description: RESTful API für die Verwaltung von Web-Appliances, Services und SSH-Verbindungen
  version: 1.0.3
  contact:
    name: Alf Lewerken
    email: support@example.com
    url: https://github.com/alflewerken/web-appliance-dashboard
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3001/api
    description: Development server
  - url: https://api.appliance-dashboard.com/api
    description: Production server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /auth/login endpoint
  schemas:
...
    AuditLog:
      type: object
      properties:
        id:
          type: integer
          example: 1234
        user_id:
          type: integer
          example: 1
        username:
          type: string
          example: "admin"
        action:
          type: string
          example: "ssh_host_updated"
        resource_type:
          type: string
          example: "ssh_host"
        resource_id:
          type: integer
          example: 1
        details:
          type: object
          properties:
            old_data:
              type: object
            new_data:
              type: object
        ip_address:
          type: string
          example: "192.168.1.100"
        created_at:
          type: string
          format: date-time
  parameters:
...
    pageParam:
      name: page
      in: query
      schema:
        type: integer
        default: 1
      description: Seitennummer
    limitParam:
      name: limit
      in: query
      schema:
        type: integer
        default: 20
      description: Einträge pro Seite
...
security:
  - bearerAuth: []
paths:
...
  /audit-logs:
    get:
      tags:
        - Audit Logs
      summary: Audit-Logs abrufen
      parameters:
        - $ref: '#/components/parameters/pageParam'
        - $ref: '#/components/parameters/limitParam'
        - name: user_id
          in: query
          schema:
            type: integer
        - name: action
          in: query
          schema:
            type: string
        - name: resource_type
          in: query
          schema:
            type: string
        - name: from
          in: query
          schema:
            type: string
            format: date-time
        - name: to
          in: query
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: Audit-Log-Liste
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  logs:
                    type: array
                    items:
                      $ref: '#/components/schemas/AuditLog'
                  pagination:
                    type: object
                    properties:
                      total:
                        type: integer
                      page:
                        type: integer
                      pages:
                        type: integer
                      limit:
                        type: integer

    delete:
      tags:
        - Audit Logs
      summary: Audit-Logs löschen
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - before
                - confirm
              properties:
                before:
                  type: string
                  format: date-time
                confirm:
                  type: boolean
      responses:
        '200':
          description: Audit-Logs gelöscht
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  deleted_count:
                    type: integer
...
  /audit:
    get:
      tags:
        - System
      summary: Audit-Logs abrufen
      parameters:
        - $ref: '#/components/parameters/pageParam'
        - $ref: '#/components/parameters/limitParam'
        - name: user_id
          in: query
          schema:
            type: integer
        - name: action
          in: query
          schema:
            type: string
        - name: from
          in: query
          schema:
            type: string
            format: date-time
        - name: to
          in: query
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: Audit-Logs
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  logs:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: integer
                        user_id:
                          type: integer
                        username:
                          type: string
                        action:
                          type: string
                        resource_type:
                          type: string
                        resource_id:
                          type: integer
                        details:
                          type: object
                        ip_address:
                          type: string
                        user_agent:
                          type: string
                        created_at:
                          type: string
                          format: date-time
                  pagination:
                    type: object
...